+++ doc/versionHistory.rst
@@ -29,6 +29,14 @@ Version History
 
 * Fix plotDonutFits to work with old aggregated data.
 
+.._lsst.ts.donut.viz-2.4.0
+
+-------------
+2.4.0
+-------------
+
+* Add unpaired aggregation tasks.
+
 .._lsst.ts.donut.viz-2.3.0
 
 -------------